requirements.txt
streamlit>=1.25.0
openai>=1.0.0
langchain>=0.1.9
langchain-community>=0.0.34
langchain-openai>=0.0.6
langchain-huggingface>=0.0.4
faiss-cpu>=1.7.4
reportlab>=4.0.0
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
pandas>=2.0.0
